refining the fix for the geolocation issue in 'listings'
