+++ src/components/listingCRUD/CreateListingPage.jsx
@@ -18,7 +18,7 @@ import ListingFacilities from "./steps/ListingFacilities";
 import ListingHouseRules from "./steps/ListingHouseRules";
 import ListingLandingPage from "./steps/ListingLandingPage";
 import ListingLocAddress from "./steps/ListingLocAddress";
-import ListingLocation from "./steps/ListingLocation";
+import ListingLocation from "./steps/listing-location/ListingLocation";
 import ListingPhotos from "./steps/ListingPhotos";
 import ListingPlaceType from "./steps/ListingPlaceType";
 import ListingPrice from "./steps/ListingPrice";
