+++ src/components/listingCRUD/steps/listing-location/react-select-styles.js
@@ -1,0 +1,19 @@
+export default {
+  container: styles => ({
+    ...styles,
+    outline: "none"
+  }),
+  input: styles => ({
+    ...styles,
+    outline: "none"
+  }),
+  control: styles => ({
+    ...styles,
+    height: "61px",
+    padding: "0 10px"
+  }),
+  indicatorSeparator: styles => ({
+    ...styles,
+    display: "none"
+  })
+};
